build fixing

+++ TapBand_SRC/Assets/Script/Controllers/MerchController.cs
@@ -8,7 +8,7 @@ public class MerchController : MonoBehaviour {
 
     public delegate void MerchTransactionEvent(MerchData merch);
     public event MerchTransactionEvent MerchTransaction;
-    public delegate void CoinTransactionEvent(int coins);
+    public delegate void CoinTransactionEvent(double coins);
     public event CoinTransactionEvent CoinTransaction;
     public delegate bool CanBuyEvent(int price);
     public event CanBuyEvent CanBuy;
